first try to save/load all kinds of posteriors
